Refined imports, extracted state models into folders and removed unnecessary declarations
