+++ src/auth/models/iauth-state.model.ts
@@ -1,0 +1,6 @@
+import { iSessionModel } from '.';
+
+export interface iAuthState {
+  session: iSessionModel | null;
+  error: string;
+}
