+++ src/auth/models/index.ts
@@ -1,1 +1,2 @@
+export * from './iauth-state.model';
 export * from './isession.model';
