+++ src/common/components/generic-form/generic-form.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { Store } from '@ngrx/store';
-import { filter, Subscription, tap } from 'rxjs';
+import { Subscription, tap } from 'rxjs';
 import {
   iGenericFormFieldModel,
   iGenericFormFieldValidation,
