+++ src/common/components/generic-modal-template/generic-modal-template.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject, TemplateRef } from '@angular/core';
-import { iGenericModalContent } from '../../models';
-import { Observable, of } from 'rxjs';
 import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { Observable, of } from 'rxjs';
+import { iGenericModalContent, iGenericModalContext } from '../../models';
 
 @Component({
   selector: 'app-generic-modal-template',
@@ -13,11 +13,11 @@ import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 export class GenericModalTemplateComponent {
   public modalContent$: Observable<iGenericModalContent | null> = of(null);
   public modalActionStatus$: Observable<boolean> = of(false);
-  public originTemplate!: TemplateRef<any>;
+  public originTemplate!: TemplateRef<Component>;
 
-  constructor(@Inject(MAT_DIALOG_DATA) data: any) {
+  constructor(@Inject(MAT_DIALOG_DATA) data: iGenericModalContext) {
     this.modalContent$ = data.modalContent$;
-    this.modalActionStatus$ = data.actionStatus$;
-    this.originTemplate = data.templateContent;
+    this.modalActionStatus$ = data.modalActionStatus$;
+    this.originTemplate = data.modalContentTemplate;
   }
 }
