+++ src/common/models/igeneric-state.model.ts
@@ -1,0 +1,19 @@
+import { iGenericModalContent, iGenericRowModel } from '.';
+
+export interface FormState {
+  isValid: boolean;
+  value: any;
+}
+
+export interface ModalState {
+  content: iGenericModalContent | null;
+  actionStatus: boolean;
+}
+
+export interface SpinnerState {
+  isLoading: boolean;
+}
+
+export interface TableState {
+  selectedRow: iGenericRowModel | null;
+}
