+++ src/common/models/index.ts
@@ -1,3 +1,4 @@
+export * from './igeneric-state.model';
 export * from './igeneric-table.model';
 export * from './igeneric-modal-content.model';
 export * from './igeneric-form-field.model';
