+++ src/common/store/actions/modal.actions.ts
@@ -1,6 +1,5 @@
 import { createAction, props } from '@ngrx/store';
 import { iGenericModalContent } from '../../models';
-import { ModalState } from '../reducers';
 
 export const resetModalState = createAction('[Modal] Reset State');
 export const setModalContent = createAction(
