+++ src/users/components/users-form/users-form.component.html
@@ -1,5 +1,4 @@
 <app-generic-form
-  #genericFormComponent
   [form]="userForm"
   [formFields]="formFields"
 ></app-generic-form>
