+++ src/users/models/index.ts
@@ -1,1 +1,2 @@
+export * from './iuser-state.model';
 export * from './iuser.model';
