+++ src/users/models/iuser-state.model.ts
@@ -1,0 +1,5 @@
+import { iUser } from '.';
+
+export interface iUsersState {
+  selectedUser: iUser | null;
+}
