+++ src/users/store/selectors/user.selector.ts
@@ -1,5 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { iUsersState } from '../users.state';
+import { iUsersState } from '../../models';
 
 export const selectUserState = createFeatureSelector<iUsersState>('users');
 export const getSelectedUser = createSelector(
